Move HTML conversion into stream instead of files
